Fixed all warnings

+++ Source/PDF Model/CPDFStream.m
@@ -46,7 +46,7 @@
     {
     CGPDFDataFormat theFormat;
     NSData *theData = (__bridge_transfer NSData *)CGPDFStreamCopyData(_stream, &theFormat);
-    return([NSString stringWithFormat:@"%@ (format: %d, length: %d)", [super description], theFormat, theData.length]);
+    return([NSString stringWithFormat:@"%@ (format: %d, length: %lu)", [super description], theFormat, (unsigned long)theData.length]);
     }
 
 - (NSData *)data
@@ -62,7 +62,7 @@
     size_t theBufferLength = strlen([thePath UTF8String]) + 1;
     char thePathBuffer[theBufferLength];
     strncpy(thePathBuffer, [thePath UTF8String], theBufferLength);
-    int theFileDescriptor = mkstemps(thePathBuffer, inPathExtension.length + 1);
+    int theFileDescriptor = mkstemps(thePathBuffer, (int)inPathExtension.length + 1);
 
     NSData *theData = self.data;
     write(theFileDescriptor, theData.bytes, inPathExtension.length + 1);
